+++ test/rpc/test_seed.py
@@ -13,7 +13,7 @@ def test_rpc_seed_threads(
     project_root, profiles_root, dbt_profile, unique_schema
 ):
     project = ProjectDefinition(
-        project_data={'seeds': {'config': {'quote_columns': False}}},
+        project_data={'seeds': {'+quote_columns': False}},
         seeds={'data.csv': 'a,b\n1,hello\n2,goodbye'},
     )
     querier_ctx = get_querier(
@@ -39,7 +39,7 @@ def test_rpc_seed_include_exclude(
     project_root, profiles_root, dbt_profile, unique_schema
 ):
     project = ProjectDefinition(
-        project_data={'seeds': {'config': {'quote_columns': False}}},
+        project_data={'seeds': {'+quote_columns': False}},
         seeds={
             'data_1.csv': 'a,b\n1,hello\n2,goodbye',
             'data_2.csv': 'a,b\n1,data',