+++ test/rpc/util.py
@@ -379,6 +379,30 @@ class Querier:
             method='test', params=params, request_id=request_id
         )
 
+    def build(
+        self,
+        models: Optional[Union[str, List[str]]] = None,
+        exclude: Optional[Union[str, List[str]]] = None,
+        threads: Optional[int] = None,
+        request_id: int = 1,
+        defer: Optional[bool] = None,
+        state: Optional[bool] = None,
+    ):
+        params = {}
+        if models is not None:
+            params['models'] = models
+        if exclude is not None:
+            params['exclude'] = exclude
+        if threads is not None:
+            params['threads'] = threads
+        if defer is not None:
+            params['defer'] = defer
+        if state is not None:
+            params['state'] = state
+        return self.request(
+            method='build', params=params, request_id=request_id
+        )
+
     def docs_generate(self, compile: bool = None, request_id: int = 1):
         params = {}
         if compile is not None: