+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## dbt 0.21.0 (Release TBD)
 
+### Under the hood
+- Add `build` RPC method, and a subset of flags for `build` task ([#3595](https://github.com/dbt-labs/dbt/issues/3595), [#3674](https://github.com/dbt-labs/dbt/pull/3674))
+
 ## dbt 0.21.0b1 (August 03, 2021)